# rubocop test: fix Performance/RedundantEqualityComparisonBlock