**Which issue(s) this PR fixes**:

  Fixes #

**What this PR does / why we need it**:

  This is cosmetic change, it does not change behavior at all. The
  following rubocop configuration detects it.

  ```
  Performance/RedundantEqualityComparisonBlock
    Enable: true
  ```

Benchmark result:

  ```
  ruby 3.2.8 (2025-03-26 revision 13f495dc2c) +YJIT [x86_64-linux]
  Warming up --------------------------------------
  all? { |t| t == XXX }
                           1.138M i/100ms
             all?(XXX)     1.330M i/100ms
  Calculating -------------------------------------
  all? { |t| t == XXX }
                           12.599M (± 1.9%) i/s   (79.37 ns/i) -     63.700M in   5.057889s
             all?(XXX)     14.230M (± 2.1%) i/s   (70.27 ns/i) -     71.821M in   5.049436s

  Comparison:
             all?(XXX): 14230128.9 i/s
  all? { |t| t == XXX }: 12598812.4 i/s - 1.13x  slower
  ```

Appendix: benchmark script:

  ```ruby
  require 'bundler/inline'
  gemfile do
    source 'https://rubygems.org'
    gem 'benchmark-ips'
  end

  Benchmark.ips do |x|
    tags = %w(apple orange grape pineapple)
    x.report("all? { |t| t == XXX }") {
      tags.all? { |t| t == "orange" }
    }
    x.report("all?(XXX)") {
      tags.all?("orange")
    }
    x.compare!
  end
  ```

**Docs Changes**:

N/A

**Release Note**: 

N/A